refactor : 기타 수정
2022. 01. 23.
- 테스트 패키지 분리
- 기타 등등

Resolves : loop-study
See also : parkeeseul
